# #1214 - Allow abstract vector in hyperrectangle fields