--- a/src/convert.jl
+++ b/src/convert.jl
@@ -460,7 +460,7 @@ A hyperrectangle.
 
 ```jldoctest
 julia> convert(Hyperrectangle, Interval(0.0, 1.0))
-Hyperrectangle{Float64}([0.5], [0.5])
+Hyperrectangle{Float64,Array{Float64,1},Array{Float64,1}}([0.5], [0.5])
 ```
 """
 function convert(::Type{Hyperrectangle}, H::AbstractHyperrectangle)
@@ -484,7 +484,7 @@ An interval.
 ### Examples
 
 ```jldoctest
-julia> convert(Interval, Hyperrectangle{Float64}([0.5], [0.5]))
+julia> convert(Interval, Hyperrectangle([0.5], [0.5]))
 Interval{Float64,IntervalArithmetic.Interval{Float64}}([0, 1])
 ```
 """