Closes #1214.