@mforets 2019-09-26
looks good and shouldn't bring performance regression.. but i would prefer that we ran benchmarks first (ReachabilityBenchmarks?)..