@schillic 2019-09-26
#### Results for `LazySets` benchmarks
`master`:

          "ρ" => 24-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "(\"LinearMap of BallInf\", \"dense\", 2)" => Trial(91.014 ns)
                  "(\"LinearMap of BallInf\", \"dense\", 1000)" => Trial(408.654 μs)
                  "(\"LinearMap of Ball1\", \"dense\", 10)" => Trial(301.405 ns)
                  "(\"Hyperrectangle\", \"dense\", 100)" => Trial(162.927 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 10)" => Trial(156.495 ns)
                  "(\"LinearMap of Ball1\", \"dense\", 100)" => Trial(4.424 μs)
                  "(\"Hyperrectangle\", \"dense\", 10)" => Trial(21.228 ns)
                  "(\"Ball1\", \"dense\", 10)" => Trial(144.747 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 100)" => Trial(3.715 μs)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 1000)" => Trial(393.918 μs)
                  "(\"BallInf\", \"dense\", 10)" => Trial(17.655 ns)
                  ⋮
          "σ" => 24-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "(\"LinearMap of BallInf\", \"dense\", 2)" => Trial(216.090 ns)
                  "(\"LinearMap of BallInf\", \"dense\", 1000)" => Trial(752.697 μs)
                  "(\"LinearMap of Ball1\", \"dense\", 10)" => Trial(420.106 ns)
                  "(\"Hyperrectangle\", \"dense\", 100)" => Trial(132.673 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 10)" => Trial(337.924 ns)
                  "(\"LinearMap of Ball1\", \"dense\", 100)" => Trial(7.643 μs)
                  "(\"Hyperrectangle\", \"dense\", 10)" => Trial(64.373 ns)
                  "(\"Ball1\", \"dense\", 10)" => Trial(126.024 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 100)" => Trial(7.005 μs)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 1000)" => Trial(755.953 μs)
                  "(\"BallInf\", \"dense\", 10)" => Trial(62.110 ns)

this branch:

          "ρ" => 24-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "(\"LinearMap of BallInf\", \"dense\", 2)" => Trial(131.000 ns)
                  "(\"LinearMap of BallInf\", \"dense\", 1000)" => Trial(444.594 μs)
                  "(\"LinearMap of Ball1\", \"dense\", 10)" => Trial(305.000 ns)
                  "(\"Hyperrectangle\", \"dense\", 100)" => Trial(190.000 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 10)" => Trial(207.000 ns)
                  "(\"LinearMap of Ball1\", \"dense\", 100)" => Trial(3.414 μs)
                  "(\"Hyperrectangle\", \"dense\", 10)" => Trial(48.000 ns)
                  "(\"Ball1\", \"dense\", 10)" => Trial(154.000 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 100)" => Trial(2.914 μs)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 1000)" => Trial(441.407 μs)
                  "(\"BallInf\", \"dense\", 10)" => Trial(45.000 ns)
                  ⋮
          "σ" => 24-element BenchmarkTools.BenchmarkGroup:
                  tags: []
                  "(\"LinearMap of BallInf\", \"dense\", 2)" => Trial(250.000 ns)
                  "(\"LinearMap of BallInf\", \"dense\", 1000)" => Trial(882.767 μs)
                  "(\"LinearMap of Ball1\", \"dense\", 10)" => Trial(418.000 ns)
                  "(\"Hyperrectangle\", \"dense\", 100)" => Trial(127.000 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 10)" => Trial(373.000 ns)
                  "(\"LinearMap of Ball1\", \"dense\", 100)" => Trial(7.931 μs)
                  "(\"Hyperrectangle\", \"dense\", 10)" => Trial(88.000 ns)
                  "(\"Ball1\", \"dense\", 10)" => Trial(139.000 ns)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 100)" => Trial(7.304 μs)
                  "(\"LinearMap of Hyperrectangle\", \"dense\", 1000)" => Trial(879.408 μs)
                  "(\"BallInf\", \"dense\", 10)" => Trial(84.000 ns)