@schillic 2019-09-26
I compared the `@code_lowered ρ(d, X)` where `X` is a `Hyperrectangle` in `master` and this branch and they are identical. So the performance changes are just noise.