@mforets 2019-09-26
the numbers returned by `PkgBenchmark` are odd, i don't expect such big differences. actually, i timed with `@benchmark` directly for Ballnf and Hyperrectangle, and can confirm that that there is no regression :+1:  . see [this notebook](https://nbviewer.jupyter.org/github/mforets/escritoire/blob/master/sets/LazySets%231717.ipynb)