@mforets 2019-09-26
> the numbers returned by PkgBenchmark are odd, i don't expect such big differences.

The command `benchmarkpkg("LazySets")` creates the file `tune.json`. By default, PkgBenchmark reuses this file from run to run if it exists. So, manually deleting this file after each call to `benchmarkpkg` allows one to recover the expected results  -- no difference between runs large than between 2% and 5%. 

On the other hand, without re-tuning the difference can go up to 100% (!) as you also found above. We can change the default behavior by passing the `retune` optional keyword argument like so: 

```julia
    results = benchmarkpkg("LazySets", retune=true)
```